+++ src/directives/index.ts
@@ -1,2 +1,3 @@
+export * from './orderable/orderable.directive';
 export * from './tableGridBodyCellTemplate/tableGridBodyCellTemplate.directive';
 export * from './tableGridHeaderCellTemplate/tableGridHeaderCellTemplate.directive';