+++ src/plugins/contentRenderer/bodyHeader/cssDivs/header/cssDivsHeaderContentRenderer.component.ts
@@ -8,6 +8,7 @@ import {HEADER_CONTENT_RENDERER_OPTIONS} from '../../../../../misc/tokens';
 import {CellContextSAPipe} from '../../../../../pipes';
 import {provideCellContextFactoryFn} from '../../../../../misc/providers';
 import {cellContextFactory} from '../../../../../misc/utils';
+import {OrderableSADirective} from '../../../../../directives';
 
 /**
  * Default options for 'CssDivsHeaderContentRendererComponent'
@@ -37,6 +38,7 @@ const defaultOptions: CssDivsHeaderContentRendererOptions =
         CommonModule,
         CellContextSAPipe,
         MergeCssClassesSAPipe,
+        OrderableSADirective,
     ],
     providers:
     [